// src/lib/proto/news.ts
export interface NewsFlash {
    /**
     * @generated from protobuf field: google.protobuf.Any any_field = 1;
     */
    anyField?: Any;
    /**
     * @generated from protobuf field: responses.NewsType type = 2;
     */
    type: NewsType;
    /**
     * @generated from protobuf field: string msg = 3;
     */
    msg: string;
}
/**
 * @generated from protobuf enum responses.NewsType
 */
export enum NewsType {
    /**
     * @generated from protobuf enum value: CHAT_TYPE_GENERIC = 0;
     */
    CHAT_TYPE_GENERIC = 0,
    /**
     * @generated from protobuf enum value: CHAT_TYPE_RESTOCK = 1;
     */
    CHAT_TYPE_RESTOCK = 1
}
// @generated message type with reflection information, may provide speed optimized methods
class NewsFlash$Type extends MessageType<NewsFlash> {
    constructor() {
        super("responses.NewsFlash", [
            { no: 1, name: "any_field", kind: "message", T: () => Any },
            { no: 2, name: "type", kind: "enum", T: () => ["responses.NewsType", NewsType] },
            { no: 3, name: "msg", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<NewsFlash>): NewsFlash {
        const message = { type: 0, msg: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<NewsFlash>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: NewsFlash): NewsFlash {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* google.protobuf.Any any_field */ 1:
                    message.anyField = Any.internalBinaryRead(reader, reader.uint32(), options, message.anyField);
                    break;
                case /* responses.NewsType type */ 2:
                    message.type = reader.int32();
                    break;
                case /* string msg */ 3:
                    message.msg = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: NewsFlash, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* google.protobuf.Any any_field = 1; */
        if (message.anyField)
            Any.internalBinaryWrite(message.anyField, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* responses.NewsType type = 2; */
        if (message.type !== 0)
            writer.tag(2, WireType.Varint).int32(message.type);
        /* string msg = 3; */
        if (message.msg !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.msg);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message responses.NewsFlash
 */
export const NewsFlash = new NewsFlash$Type();